requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.2
cohere>=4.45
faiss-cpu>=1.7.4
python-docx>=1.0.1
PyPDF2>=3.0.0
tiktoken>=0.5.2
tenacity>=8.2.3
anthropic>=0.18.1 